C# code
This program has the ability to generate an Arithmetic Progression using for loop and user defined a1 - first element; n - number of elements; d - the difference between elements. The program also has the ability to determine the first and the last element and successfully calculate the sum of the progression.